Implement auth guards

+++ src/leaves/leaves.service.ts
@@ -12,32 +12,49 @@ export class LeavesService {
     private leaveRepository: LeaveRepository,
   ) {}
 
-  async getAllLeaves(): Promise<Leave[]> {
-    return await this.leaveRepository.find();
+  async getAllLeaves(userId: string): Promise<Leave[]> {
+    return await this.leaveRepository.find({
+      where: {
+        userId: {
+          $eq: userId,
+        },
+      },
+    });
   }
 
-  async getLeaveById(id: string): Promise<Leave> {
-    return this.leaveRepository.getLeaveById(id);
+  async getLeaveById(id: string, userId: string): Promise<Leave> {
+    return this.leaveRepository.getLeaveById(id, userId);
   }
 
-  async createLeave(leaveDatesDto: LeaveDatesDto): Promise<Leave> {
-    return this.leaveRepository.createLeave(leaveDatesDto);
+  async createLeave(
+    leaveDatesDto: LeaveDatesDto,
+    userId: string,
+  ): Promise<Leave> {
+    return this.leaveRepository.createLeave(leaveDatesDto, userId);
   }
 
-  async updateLeaveStatus(id: string, status: LeaveStatus): Promise<Leave> {
-    const leave = await this.getLeaveById(id);
-    leave.status = status;
-    return await this.leaveRepository.save(leave);
+  async updateLeaveStatus(
+    id: string,
+    status: LeaveStatus,
+    userId: string,
+  ): Promise<Leave> {
+    const leave = await this.getLeaveById(id, userId);
+    if (leave.status !== status) {
+      leave.status = status;
+      return await this.leaveRepository.save(leave);
+    }
+    return leave;
   }
 
   async updateLeaveDates(
     id: string,
     leaveDatesDto: LeaveDatesDto,
+    userId: string,
   ): Promise<Leave> {
-    return this.leaveRepository.updateLeaveDates(id, leaveDatesDto);
+    return this.leaveRepository.updateLeaveDates(id, leaveDatesDto, userId);
   }
 
-  async deleteLeave(id: string): Promise<void> {
-    this.leaveRepository.delete({ id });
+  async deleteLeave(id: string, userId: string): Promise<void> {
+    this.leaveRepository.delete({ id, userId });
   }
 }
